requirements.txt
groq>=0.9.0
streamlit>=1.39.0
duckduckgo-search>=6.2.0
wikipedia>=1.4.0
requests>=2.31.0
python-dotenv>=1.0.0